breaking up meal card
